+++ apps/ledger-live-mobile/src/components/FirebaseFeatureFlags.tsx
@@ -1,12 +1,15 @@
 import React, { ReactNode } from "react";
+import { useSelector } from "react-redux";
 import remoteConfig from "@react-native-firebase/remote-config";
 import {
   FeatureFlagsProvider,
   defaultFeatures,
 } from "@ledgerhq/live-common/featureFlags/index";
-import { FeatureId } from "@ledgerhq/live-common/types/index";
+import { FeatureId } from "@ledgerhq/types-live";
 
 import { formatFeatureId } from "./FirebaseRemoteConfig";
+
+import { languageSelector } from "../reducers/settings";
 
 type Props = {
   children?: ReactNode;
@@ -16,6 +19,19 @@ const getFeature = (key: FeatureId) => {
   try {
     const value = remoteConfig().getValue(formatFeatureId(key));
     const feature = JSON.parse(value.asString());
+    const currAppLanguage = useSelector(languageSelector);
+
+    if (
+      (feature.languages_whitelisted &&
+        !feature.languages_whitelisted.includes(currAppLanguage)) ||
+      (feature.languages_blacklisted &&
+        feature.languages_blacklisted.includes(currAppLanguage))
+    ) {
+      return {
+        ...feature,
+        enabled: false,
+      };
+    }
 
     return feature;
   } catch (error) {